+++ src/main/js/plugin/input-binding/color-number-input.ts
@@ -1,5 +1,6 @@
 import {InputParams} from '../../api/types';
 import {InputBinding} from '../../binding/input';
+import {ColorSwatchTextInputController} from '../../controller/input/color-swatch-text';
 import * as ColorConverter from '../../converter/color';
 import {ColorFormatter} from '../../formatter/color';
 import {Color} from '../../model/color';
@@ -7,7 +8,6 @@ import {InputValue} from '../../model/input-value';
 import {ViewModel} from '../../model/view-model';
 import * as NumberColorParser from '../../parser/number-color';
 import * as StringColorParser from '../../parser/string-color';
-import {ColorSwatchTextInputController} from '../input/color-swatch-text';
 import {InputBindingPlugin} from './input-binding-plugin';
 
 function shouldSupportAlpha(inputParams: InputParams): boolean {
