+++ src/main/js/plugin/input-binding/number-input.ts
@@ -6,16 +6,16 @@ import {ListConstraint} from '../../constraint/list';
 import {RangeConstraint} from '../../constraint/range';
 import {StepConstraint} from '../../constraint/step';
 import {ConstraintUtil} from '../../constraint/util';
+import {ListInputController} from '../../controller/input/list';
+import {NumberTextInputController} from '../../controller/input/number-text';
+import {SliderTextInputController} from '../../controller/input/slider-text';
+import * as UiUtil from '../../controller/ui-util';
 import * as NumberConverter from '../../converter/number';
 import {NumberFormatter} from '../../formatter/number';
 import {TypeUtil} from '../../misc/type-util';
 import {InputValue} from '../../model/input-value';
 import {ViewModel} from '../../model/view-model';
 import {StringNumberParser} from '../../parser/string-number';
-import {ListInputController} from '../input/list';
-import {NumberTextInputController} from '../input/number-text';
-import {SliderTextInputController} from '../input/slider-text';
-import * as UiUtil from '../ui-util';
 import {InputBindingPlugin} from './input-binding-plugin';
 
 function createConstraint(params: InputParams): Constraint<number> {
